ANDI_to_SR decoder

+++ src/devices/CPU/include/cpu/internal/instructions/data/ANDI_to_SR_instruction_data.h
@@ -4,7 +4,7 @@
 namespace m68k::InstructionData {
 
 struct ANDI_to_SR_InstructionData {
-    uint8_t immediateData;
+    uint16_t immediateData;
 };    
 
 } // namespace m68k::InstructionData
